# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Trademark Ledger & Case Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /cases:
    get:
      operationId: listCases
      tags: [cases]
      summary: List / search trademark cases
      parameters:
        - name: folderNumber
          in: query
          schema: { type: string }
        - name: tmNumber
          in: query
          schema: { type: string }
        - name: clientNumber
          in: query
          schema: { type: string }
        - name: applicantName
          in: query
          schema: { type: string }
        - name: class
          in: query
          schema: { type: string }
        - name: stage
          in: query
          schema: { type: integer }
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
      responses:
        "200":
          description: List of cases
          content:
            application/json:
              schema:
                type: object
                required: [data, total]
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/TrademarkCase" }
                  total: { type: integer }
    post:
      operationId: createCase
      tags: [cases]
      summary: Create a trademark case
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CaseInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TrademarkCase" }
  /cases/{folderId}:
    get:
      operationId: getCase
      tags: [cases]
      summary: Get a case by folder number
      parameters:
        - name: folderId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Case detail
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TrademarkCaseDetail" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
    patch:
      operationId: updateCase
      tags: [cases]
      summary: Update a trademark case
      parameters:
        - name: folderId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CaseUpdate" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TrademarkCase" }
    delete:
      operationId: deleteCase
      tags: [cases]
      summary: Delete a trademark case
      parameters:
        - name: folderId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DeletedResponse" }

  # ── STAGES ───────────────────────────────────────────────────────────────
...
  /cases/{folderId}/assignments:
    get:
      operationId: listAssignments
      tags: [assignments]
      summary: List all Stage 2 assignments for a case
      parameters:
        - name: folderId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Assignments
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Assignment" }
    post:
      operationId: createAssignment
      tags: [assignments]
      summary: Assign a person to Stage 2
      parameters:
        - name: folderId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AssignmentInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Assignment" }
  /cases/{folderId}/assignments/{assignmentId}:
    patch:
      operationId: updateAssignment
      tags: [assignments]
      summary: Update an assignment (accept date, status, etc.)
      parameters:
        - name: folderId
          in: path
          required: true
          schema: { type: string }
        - name: assignmentId
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AssignmentUpdate" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Assignment" }
    delete:
      operationId: deleteAssignment
      tags: [assignments]
      summary: Delete an assignment
      parameters:
        - name: folderId
          in: path
          required: true
          schema: { type: string }
        - name: assignmentId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DeletedResponse" }

  # ── CASE LEDGER ──────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error: { type: string }
    DeletedResponse:
      type: object
      required: [success]
      properties:
        success: { type: boolean }

    # ── CLIENT ──────────────────────────────────────────────────────────────
...
    TrademarkCase:
      type: object
      required: [id, folderNumber, clientId, applicantName, stage, createdAt, updatedAt]
      properties:
        id: { type: integer }
        folderNumber: { type: string }
        clientId: { type: integer }
        clientNumber: { type: ["string", "null"] }
        clientName: { type: ["string", "null"] }
        tmNumber: { type: ["string", "null"] }
        applicantName: { type: string }
        class: { type: ["string", "null"] }
        filingDate: { type: ["string", "null"] }
        stage: { type: integer }
        subStage: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
        createdAt: { type: string }
        updatedAt: { type: string }
    TrademarkCaseDetail:
      type: object
      required: [id, folderNumber, clientId, applicantName, stage, stages, createdAt, updatedAt]
      properties:
        id: { type: integer }
        folderNumber: { type: string }
        clientId: { type: integer }
        clientNumber: { type: ["string", "null"] }
        clientName: { type: ["string", "null"] }
        tmNumber: { type: ["string", "null"] }
        applicantName: { type: string }
        class: { type: ["string", "null"] }
        filingDate: { type: ["string", "null"] }
        stage: { type: integer }
        subStage: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
        stages:
          type: array
          items: { $ref: "#/components/schemas/CaseStage" }
        assignments:
          type: array
          items: { $ref: "#/components/schemas/Assignment" }
        createdAt: { type: string }
        updatedAt: { type: string }
    CaseInput:
      type: object
      required: [folderNumber, clientId, applicantName]
      properties:
        folderNumber: { type: string }
        clientId: { type: integer }
        tmNumber: { type: string }
        applicantName: { type: string }
        class: { type: string }
        filingDate: { type: string }
        stage: { type: integer, default: 1 }
        subStage: { type: string }
        notes: { type: string }
    CaseUpdate:
      type: object
      properties:
        tmNumber: { type: string }
        applicantName: { type: string }
        class: { type: string }
        filingDate: { type: string }
        stage: { type: integer }
        subStage: { type: string }
        notes: { type: string }

    # ── STAGE ────────────────────────────────────────────────────────────────
    StageTimelineEvent:
      type: object
      required: [date, description]
      properties:
        date: { type: string }
        description: { type: string }
    StageHistoryEntry:
      type: object
      required: [changedAt, field, oldValue, newValue]
      properties:
        changedAt: { type: string }
        field: { type: string }
        oldValue: { type: ["string", "null"] }
        newValue: { type: ["string", "null"] }
    CaseStage:
      type: object
      required: [id, caseId, stageNumber, createdAt, updatedAt]
      properties:
        id: { type: integer }
        caseId: { type: integer }
        stageNumber: { type: integer }
        status: { type: ["string", "null"] }
        subStatus: { type: ["string", "null"] }
        timeline:
          type: array
          items: { $ref: "#/components/schemas/StageTimelineEvent" }
        history:
          type: array
          items: { $ref: "#/components/schemas/StageHistoryEntry" }
        createdAt: { type: string }
        updatedAt: { type: string }
...
    Assignment:
      type: object
      required: [id, caseId, personName, assignedDate, status, createdAt]
      properties:
        id: { type: integer }
        caseId: { type: integer }
        personName: { type: string }
        city: { type: ["string", "null"] }
        assignedDate: { type: string }
        acceptedDate: { type: ["string", "null"] }
        status: { type: string }
        createdAt: { type: string }
        updatedAt: { type: string }
    AssignmentInput:
      type: object
      required: [personName, assignedDate]
      properties:
        personName: { type: string }
        city: { type: string }
        assignedDate: { type: string }
        status: { type: string, default: "Pending" }
    AssignmentUpdate:
      type: object
      properties:
        personName: { type: string }
        city: { type: string }
        assignedDate: { type: string }
        acceptedDate: { type: string }
        status: { type: string }

    # ── LEDGER ───────────────────────────────────────────────────────────────
...